Rearrange tests
